Bulk Upload Release Fixes

+++ pipeline/alembic/versions/4000b95a31a2_imported_ia_add_group_id.py
@@ -7,6 +7,7 @@ Create Date: 2020-08-10 19:07:26.929935
 """
 from alembic import op
 import sqlalchemy as sa
+from sqlalchemy.engine.reflection import Inspector
 
 
 # revision identifiers, used by Alembic.
@@ -21,4 +22,9 @@ def upgrade():
 
 
 def downgrade():
-    op.drop_column("imported_industrial_athlete", "group_id")
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
+    
+    if 'imported_industrial_athlete' in tables:
+        op.drop_column("imported_industrial_athlete", "group_id")
