+++ pipeline/alembic/versions/ace2c9349ddf_populate_cabin_rule_table.py
@@ -9,17 +9,11 @@ from alembic import op
 import sqlalchemy as sa
 from sqlalchemy import orm
 from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.engine.reflection import Inspector
 from sqlalchemy.exc import InternalError
 
 RBAC_VALID_RESOURCES = (
     "athletes",
-    "clients",
-    "shifts",
-    "docks",
-    "warehouses",
-    "jobfunctions",
-    "roles",
-    "bulkupload",
 )
 
 RBAC_ACTION_KEYS = {"get": "read", "post": "write", "put": "update", "delete": "delete"}
@@ -51,16 +45,15 @@ def build_records():
     records = []
     for resource in RBAC_VALID_RESOURCES:
         for action in list(RBAC_ACTION_KEYS.keys()):
-            if resource in ("athletes", "bulkupload"):
+            if resource in ("athletes"):
                 records.append(CasbinRule(v0="manager", v1=resource, v2=action))
             records.append(CasbinRule(v0="admin", v1=resource, v2=action))
     return records
 
-
 def upgrade():
     bind = op.get_bind()
     session = orm.Session(bind=bind)
-    records = build_records()
+    #records = build_records()
     try:
         CasbinRule.__table__.create(bind)
     except InternalError as e:
@@ -68,10 +61,15 @@ def upgrade():
         if code == "1050":
             session.query(CasbinRule).delete()
     finally:
-        session.add_all(records)
-        session.commit()
+        #session.add_all(records)
+        #session.commit()
         session.close()
 
 
 def downgrade():
-    op.drop_table('casbin_rule')
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
+
+    if 'casbin_rule' in tables:
+        op.drop_table("casbin_rule")
