+++ pipeline/alembic/versions/b481cf862c87_delete_athlete_upload_status_id.py
@@ -6,6 +6,7 @@ Create Date: 2020-08-21 18:09:06.629114
 
 """
 from alembic import op
+from sqlalchemy.engine.reflection import Inspector
 import sqlalchemy as sa
 
 
@@ -17,7 +18,12 @@ depends_on = None
 
 
 def upgrade():
-    op.drop_table("imported_industrial_athlete")
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
+    
+    if 'imported_industrial_athlete' in tables:
+        op.drop_table("imported_industrial_athlete")
 
 
 def downgrade():
