+++ pipeline/alembic/versions/e175652bbce6_add_athlete_upload_status_id.py
@@ -7,6 +7,7 @@ Create Date: 2020-08-21 18:09:19.502570
 """
 import datetime
 from alembic import op
+from sqlalchemy.engine.reflection import Inspector
 import sqlalchemy as sa
 
 
@@ -59,5 +60,9 @@ def upgrade():
 
 
 def downgrade():
-    op.drop_table("imported_industrial_athlete")
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
 
+    if 'imported_industrial_athlete' in tables:
+        op.drop_table("imported_industrial_athlete")
